qt: split gamelist into two seperate files
